Made command line arguments
Added a new Main so that it would take two arguments for testing, one
file that we want to shift and one that we are trying to shift it to.
Files should be in music test folder. Still trying to figure out a
better way to implement autotune, but for now that's all I got.